[fix] 함수 분리 및 playedPlayer 조건문 수정
