+++ src/components/features/player/DetailPlayer.tsx
@@ -1,5 +1,5 @@
 import { getYoutubeID } from "@/utils/getYoutubeID";
-import { getSpotifyTrack } from "@/utils/supabase/server-actions";
+import { getSpotifyTrack } from "@/utils/spotify-actions";
 import DetailPlayButton from "./DetailPlayButton";
 
 type Props = {
